Int 730 update tenant config end points to be consistent (#132)
* Updates to support new tenant config in proxy server

* Tenant service updates

* Lint updates

* More lint updates

* Updates to support patientMRNTypeText

* Updates to support tenant and ehr updates with only an object

+++ interop-tenant/src/main/kotlin/com/projectronin/interop/tenant/config/data/binding/ProviderPoolDOs.kt
@@ -3,11 +3,10 @@ package com.projectronin.interop.tenant.config.data.binding
 import com.projectronin.interop.tenant.config.data.model.ProviderPoolDO
 import org.ktorm.schema.Table
 import org.ktorm.schema.int
-import org.ktorm.schema.long
 import org.ktorm.schema.varchar
 
 object ProviderPoolDOs : Table<ProviderPoolDO>("io_tenant_provider_pool") {
-    val id = long("io_tenant_provider_pool_id").primaryKey().bindTo { it.id }
+    val id = int("io_tenant_provider_pool_id").primaryKey().bindTo { it.id }
     val tenantId = int("io_tenant_id").references(TenantDOs) { it.tenant }
     val providerId = varchar("provider_id").bindTo { it.providerId }
     val poolId = varchar("pool_id").bindTo { it.poolId }
